refactor: apply copilot feedback on template route helpers
